+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 1. ARM_FE - Designing and Implementing Student Profile creation.
 2. ARM_FE - Designing and Implementing Event Creation.
+2. ARM_FE - Designing and Implementing Student Event Search by Date and Roll No.